do not drop index

+++ crf-migration/src/test/java/utils/MongoDBUtils.java
@@ -50,6 +50,7 @@ public class MongoDBUtils {
                                                   migratorConfig.getSourceMongoConfig().getPassword());
 
             for (String name: sourceMongo.getDB(migratorConfig.getSourceMongoConfig().getDbName()).getCollectionNames()) {
+                if ("system.indexes".equalsIgnoreCase(name)) continue;
                 sourceMongo.getDB(migratorConfig.getTargetMongoConfig().getDbName())
                            .getCollection(name).drop();
             }
@@ -63,6 +64,7 @@ public class MongoDBUtils {
                                                  );
 
             for (String name: targetMongo.getDB(migratorConfig.getTargetMongoConfig().getDbName()).getCollectionNames()) {
+                if ("system.indexes".equalsIgnoreCase(name)) continue;
                 targetMongo.getDB(migratorConfig.getTargetMongoConfig().getDbName())
                            .getCollection(name).drop();
             }
